Extract files route controller
