+++ js/etherpad_nextcloud-viewer-main.mjs
@@ -1,2 +1,2 @@
-import{V as g,M as b,t as p,a as f,A as y,o as m}from"./oc-compat-DDPKm7Cs.chunk.mjs";import{a as L,e as I}from"./urls-B3dfoIpo.chunk.mjs";(function(){let v=0;const S={name:"EtherpadNextcloudViewer",props:{filename:{type:String,required:!1,default:""},basename:{type:String,required:!1,default:""},source:{type:String,required:!1,default:""},fileid:{type:[String,Number],required:!1,default:null},fileId:{type:[String,Number],required:!1,default:null},fileInfo:{type:Object,required:!1,default:null}},data(){return{iframeSrc:"",isLoading:!0,loadError:"",externalOpenUrl:"",externalOpenMessage:"",externalSnapshotText:"",readonlySnapshotMode:!1,readonlySnapshotText:"",readonlySnapshotHtml:"",resolveGeneration:0,syncUrl:"",syncIntervalMs:12e4,syncInFlight:!1,syncTimerId:null,visibilityHandler:null,pageHideHandler:null}},computed:{sourcePath(){const e=typeof this.source=="string"?this.source.trim():"";return e&&I(e)||""},filePath(){const e=t=>String(t||"").trim().replace(/\s+\.pad$/i,".pad"),n=t=>{const r=String(t||"").trim();return!r||r==="/"?"/":r.startsWith("/")?r:"/"+r},s=(t,r)=>{if(!r)return"";if(r.startsWith("/"))return r;const o=n(t);return o==="/"?"/"+r:o+"/"+r},a=t=>typeof t=="string"&&t.toLowerCase().endsWith(".pad");if(a(this.sourcePath))return this.sourcePath;const i=this.fileInfo&&typeof this.fileInfo=="object"?this.fileInfo:null,h=i&&typeof i.path=="string"?e(i.path):"";if(a(h))return h.startsWith("/")?h:"/"+h;const d=e(this.filename||this.basename||i&&(i.name||i.basename)||"");if(!d)return"";const c=i&&typeof i.dirname=="string"?i.dirname:"";if(c){const t=s(c,d);if(a(t))return t}const u=new URLSearchParams(window.location.search||"").get("dir")||"/",l=s(u,d);return a(l)?l:"/"+d},resolvedFileId(){const e=[this.fileid,this.fileId,this.fileInfo&&(this.fileInfo.fileid||this.fileInfo.fileId||this.fileInfo.id)];for(const a of e){const i=Number(a);if(Number.isFinite(i)&&i>0)return i}if(new URLSearchParams(window.location.search||"").get("openfile")!=="true")return null;const n=(window.location.pathname||"").match(/\/apps\/files\/files\/(\d+)\/?$/);if(!n)return null;const s=Number(n[1]);return Number.isFinite(s)&&s>0?s:null}},watch:{filePath:{immediate:!0,handler(){this.resolveOpenUrl()}},resolvedFileId(){this.resolveOpenUrl()}},methods:{async fetchOpenPayload(e,n={}){const s=Object.assign({Accept:"application/json"},n.headers||{}),a=await fetch(e,Object.assign({method:"GET",credentials:"same-origin",headers:s},n)),i=await a.json().catch(()=>({}));if(!a.ok)throw new Error(i&&i.message||"Pad open failed.");if(!i||i.is_readonly_snapshot!==!0&&(typeof i.url!="string"||i.url.trim()===""))throw new Error("Pad open API did not return a valid URL.");return i},isMissingFrontmatterError(e){return e instanceof Error?String(e.message||"").includes("Missing YAML frontmatter"):!1},async initializeMissingFrontmatter(){const e={Accept:"application/json",requesttoken:m()};if(this.resolvedFileId!==null){const h=f("/apps/"+y+"/api/v1/pads/initialize-by-id/"+encodeURIComponent(String(this.resolvedFileId))),d=await fetch(h,{method:"POST",credentials:"same-origin",headers:e}),c=await d.json().catch(()=>({}));if(!d.ok)throw new Error(c&&c.message||"Pad initialization failed.");return c}if(!this.filePath)throw new Error("Pad initialization failed: missing file path.");const n=new URLSearchParams;n.set("file",this.filePath);const s=f("/apps/"+y+"/api/v1/pads/initialize"),a=await fetch(s,{method:"POST",credentials:"same-origin",headers:Object.assign({},e,{"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8"}),body:n.toString()}),i=await a.json().catch(()=>({}));if(!a.ok)throw new Error(i&&i.message||"Pad initialization failed.");return i},markLoaded(){this.$emit("update:loaded",!0)},stopSyncLoop(){this.syncTimerId!==null&&(window.clearInterval(this.syncTimerId),this.syncTimerId=null)},startSyncLoop(){!this.syncUrl||this.syncTimerId!==null||(this.syncTimerId=window.setInterval(()=>{document.visibilityState==="visible"&&this.runSync(!1,!1)},this.syncIntervalMs))},async runSync(e,n){if(this.syncUrl&&!(this.syncInFlight&&!e)){this.syncInFlight=!0;try{const s=e?this.syncUrl+(this.syncUrl.includes("?")?"&":"?")+"force=1":this.syncUrl;await fetch(s,{method:"POST",credentials:"same-origin",headers:{Accept:"application/json",requesttoken:m()},keepalive:!!n})}finally{this.syncInFlight=!1}}},installSyncLifecycleHandlers(){this.visibilityHandler||this.pageHideHandler||(this.visibilityHandler=()=>{if(document.visibilityState==="hidden"){this.runSync(!0,!0),this.stopSyncLoop();return}this.startSyncLoop()},this.pageHideHandler=()=>{this.runSync(!0,!0),this.stopSyncLoop()},document.addEventListener("visibilitychange",this.visibilityHandler),window.addEventListener("pagehide",this.pageHideHandler))},removeSyncLifecycleHandlers(){this.visibilityHandler&&(document.removeEventListener("visibilitychange",this.visibilityHandler),this.visibilityHandler=null),this.pageHideHandler&&(window.removeEventListener("pagehide",this.pageHideHandler),this.pageHideHandler=null)},async resolveOpenUrl(){const e=++this.resolveGeneration,n=()=>e===this.resolveGeneration;if(this.isLoading=!0,this.loadError="",this.iframeSrc="",this.externalOpenUrl="",this.externalOpenMessage="",this.externalSnapshotText="",this.readonlySnapshotMode=!1,this.readonlySnapshotText="",this.readonlySnapshotHtml="",this.syncUrl="",this.syncInFlight=!1,this.stopSyncLoop(),!this.filePath){if(!n())return;this.loadError="No .pad file selected.",this.isLoading=!1;return}const s=L(),a=f("/apps/"+y+"/api/v1/pads/open"),i=(()=>{if(!s)return"";const l=new URL(f("/apps/"+y+"/api/v1/public/open/"+encodeURIComponent(s)),window.location.origin);return l.searchParams.set("file",this.filePath),l.toString()})(),h=this.resolvedFileId!==null?f("/apps/"+y+"/api/v1/pads/open-by-id"):"",d=new URLSearchParams;d.set("file",this.filePath);const c=new URLSearchParams;this.resolvedFileId!==null&&c.set("fileId",String(this.resolvedFileId));const u={"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:m()};try{const l=async()=>{let o=null;if(i)o=await this.fetchOpenPayload(i);else if(h)try{o=await this.fetchOpenPayload(h,{method:"POST",headers:u,body:c.toString()})}catch{}return o||(o=await this.fetchOpenPayload(a,{method:"POST",headers:u,body:d.toString()})),o};let t;try{if(t=await l(),!n())return}catch(o){if(!this.isMissingFrontmatterError(o))throw o;if(await this.initializeMissingFrontmatter(),!n()||(t=await l(),!n()))return}this.syncUrl=t&&typeof t.sync_url=="string"?t.sync_url:"";const r=Number(t&&t.sync_interval_seconds);if(this.syncIntervalMs=Number.isFinite(r)&&r>0?Math.max(5e3,Math.min(36e5,r*1e3)):12e4,this.installSyncLifecycleHandlers(),this.syncUrl&&this.startSyncLoop(),t&&t.is_readonly_snapshot===!0){this.readonlySnapshotMode=!0,this.readonlySnapshotText=typeof t.snapshot_text=="string"?t.snapshot_text:"",this.readonlySnapshotHtml=typeof t.snapshot_html=="string"?t.snapshot_html:"",this.markLoaded();return}if(t&&t.is_external===!0&&typeof t.url=="string"&&t.url.trim()!==""){const o=t.url.trim();this.externalOpenUrl=o,this.externalOpenMessage=p("This view shows the last synced snapshot stored in the .pad file. It is read-only here. To edit the pad, open the original pad in a new tab."),this.externalSnapshotText=t&&typeof t.snapshot_text=="string"?t.snapshot_text:"",this.markLoaded();return}this.iframeSrc=t.url,this.markLoaded()}catch(l){if(!n())return;this.loadError=l instanceof Error?l.message:"Could not load pad.",this.markLoaded()}finally{if(!n())return;this.isLoading=!1}}},beforeDestroy(){this.resolveGeneration+=1,this.runSync(!0,!0),this.stopSyncLoop(),this.removeSyncLifecycleHandlers()},beforeUnmount(){this.resolveGeneration+=1,this.runSync(!0,!0),this.stopSyncLoop(),this.removeSyncLifecycleHandlers()},render(e){return this.loadError?e("div",{class:"epnc-native-status epnc-native-status--error"},[e("div",{class:"epnc-native-error-card"},[e("div",{class:"epnc-native-error-title"},"Unable to open pad"),e("div",{class:"epnc-native-error-message"},this.loadError)])]):this.externalOpenUrl?e("div",{class:"epnc-native-status"},[e("div",{class:"epnc-native-error-card"},[e("div",{class:"epnc-native-error-title"},p("Pad from another server")),e("div",{class:"epnc-native-error-message"},this.externalOpenMessage),e("pre",{class:"epnc-native-preview"},this.externalSnapshotText.trim()!==""?this.externalSnapshotText:p("No synced snapshot is stored in this .pad file yet.")),e("a",{class:"button primary",attrs:{href:this.externalOpenUrl,target:"_blank",rel:"noopener noreferrer"}},p("Open original pad in new tab"))])]):this.readonlySnapshotMode?e("div",{class:"epnc-native-snapshot"},[e("div",{class:"epnc-native-snapshot__inner"},[e("div",{class:"epnc-native-snapshot__title"},p("Read-only snapshot")),e("div",{class:"epnc-native-snapshot__message"},p("This share shows the last synced snapshot stored in the .pad file.")),this.readonlySnapshotHtml.trim()!==""?e("div",{class:"epnc-native-snapshot__text epnc-native-snapshot__text--html",domProps:{innerHTML:this.readonlySnapshotHtml}}):e("pre",{class:"epnc-native-snapshot__text"},this.readonlySnapshotText.trim()!==""?this.readonlySnapshotText:p("No synced snapshot is stored in this .pad file yet."))])]):this.isLoading||!this.iframeSrc?e("div",{class:"epnc-native-status"},"Loading pad..."):e("div",{class:"epnc-native-shell"},[e("iframe",{attrs:{src:this.iframeSrc},on:{load:()=>this.markLoaded(),error:()=>this.markLoaded()},class:"epnc-native-iframe"})])}},w=()=>{if(v+=1,!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.registerHandler=="function")){v<20&&window.setTimeout(w,500);return}Array.isArray(window.OCA.Viewer.availableHandlers)&&window.OCA.Viewer.availableHandlers.some(e=>e&&e.id===g)||window.OCA.Viewer.registerHandler({id:g,mimes:[b],component:S})};w()})();
+import{V as g,M as b,t as p,c as f,A as y,o as m}from"./oc-compat-C8UpNj_V.chunk.mjs";import{a as L,e as I}from"./urls-M9FeF2zV.chunk.mjs";(function(){let v=0;const S={name:"EtherpadNextcloudViewer",props:{filename:{type:String,required:!1,default:""},basename:{type:String,required:!1,default:""},source:{type:String,required:!1,default:""},fileid:{type:[String,Number],required:!1,default:null},fileId:{type:[String,Number],required:!1,default:null},fileInfo:{type:Object,required:!1,default:null}},data(){return{iframeSrc:"",isLoading:!0,loadError:"",externalOpenUrl:"",externalOpenMessage:"",externalSnapshotText:"",readonlySnapshotMode:!1,readonlySnapshotText:"",readonlySnapshotHtml:"",resolveGeneration:0,syncUrl:"",syncIntervalMs:12e4,syncInFlight:!1,syncTimerId:null,visibilityHandler:null,pageHideHandler:null}},computed:{sourcePath(){const e=typeof this.source=="string"?this.source.trim():"";return e&&I(e)||""},filePath(){const e=t=>String(t||"").trim().replace(/\s+\.pad$/i,".pad"),n=t=>{const r=String(t||"").trim();return!r||r==="/"?"/":r.startsWith("/")?r:"/"+r},s=(t,r)=>{if(!r)return"";if(r.startsWith("/"))return r;const o=n(t);return o==="/"?"/"+r:o+"/"+r},a=t=>typeof t=="string"&&t.toLowerCase().endsWith(".pad");if(a(this.sourcePath))return this.sourcePath;const i=this.fileInfo&&typeof this.fileInfo=="object"?this.fileInfo:null,h=i&&typeof i.path=="string"?e(i.path):"";if(a(h))return h.startsWith("/")?h:"/"+h;const d=e(this.filename||this.basename||i&&(i.name||i.basename)||"");if(!d)return"";const c=i&&typeof i.dirname=="string"?i.dirname:"";if(c){const t=s(c,d);if(a(t))return t}const u=new URLSearchParams(window.location.search||"").get("dir")||"/",l=s(u,d);return a(l)?l:"/"+d},resolvedFileId(){const e=[this.fileid,this.fileId,this.fileInfo&&(this.fileInfo.fileid||this.fileInfo.fileId||this.fileInfo.id)];for(const a of e){const i=Number(a);if(Number.isFinite(i)&&i>0)return i}if(new URLSearchParams(window.location.search||"").get("openfile")!=="true")return null;const n=(window.location.pathname||"").match(/\/apps\/files\/files\/(\d+)\/?$/);if(!n)return null;const s=Number(n[1]);return Number.isFinite(s)&&s>0?s:null}},watch:{filePath:{immediate:!0,handler(){this.resolveOpenUrl()}},resolvedFileId(){this.resolveOpenUrl()}},methods:{async fetchOpenPayload(e,n={}){const s=Object.assign({Accept:"application/json"},n.headers||{}),a=await fetch(e,Object.assign({method:"GET",credentials:"same-origin",headers:s},n)),i=await a.json().catch(()=>({}));if(!a.ok)throw new Error(i&&i.message||"Pad open failed.");if(!i||i.is_readonly_snapshot!==!0&&(typeof i.url!="string"||i.url.trim()===""))throw new Error("Pad open API did not return a valid URL.");return i},isMissingFrontmatterError(e){return e instanceof Error?String(e.message||"").includes("Missing YAML frontmatter"):!1},async initializeMissingFrontmatter(){const e={Accept:"application/json",requesttoken:m()};if(this.resolvedFileId!==null){const h=f("/apps/"+y+"/api/v1/pads/initialize-by-id/"+encodeURIComponent(String(this.resolvedFileId))),d=await fetch(h,{method:"POST",credentials:"same-origin",headers:e}),c=await d.json().catch(()=>({}));if(!d.ok)throw new Error(c&&c.message||"Pad initialization failed.");return c}if(!this.filePath)throw new Error("Pad initialization failed: missing file path.");const n=new URLSearchParams;n.set("file",this.filePath);const s=f("/apps/"+y+"/api/v1/pads/initialize"),a=await fetch(s,{method:"POST",credentials:"same-origin",headers:Object.assign({},e,{"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8"}),body:n.toString()}),i=await a.json().catch(()=>({}));if(!a.ok)throw new Error(i&&i.message||"Pad initialization failed.");return i},markLoaded(){this.$emit("update:loaded",!0)},stopSyncLoop(){this.syncTimerId!==null&&(window.clearInterval(this.syncTimerId),this.syncTimerId=null)},startSyncLoop(){!this.syncUrl||this.syncTimerId!==null||(this.syncTimerId=window.setInterval(()=>{document.visibilityState==="visible"&&this.runSync(!1,!1)},this.syncIntervalMs))},async runSync(e,n){if(this.syncUrl&&!(this.syncInFlight&&!e)){this.syncInFlight=!0;try{const s=e?this.syncUrl+(this.syncUrl.includes("?")?"&":"?")+"force=1":this.syncUrl;await fetch(s,{method:"POST",credentials:"same-origin",headers:{Accept:"application/json",requesttoken:m()},keepalive:!!n})}finally{this.syncInFlight=!1}}},installSyncLifecycleHandlers(){this.visibilityHandler||this.pageHideHandler||(this.visibilityHandler=()=>{if(document.visibilityState==="hidden"){this.runSync(!0,!0),this.stopSyncLoop();return}this.startSyncLoop()},this.pageHideHandler=()=>{this.runSync(!0,!0),this.stopSyncLoop()},document.addEventListener("visibilitychange",this.visibilityHandler),window.addEventListener("pagehide",this.pageHideHandler))},removeSyncLifecycleHandlers(){this.visibilityHandler&&(document.removeEventListener("visibilitychange",this.visibilityHandler),this.visibilityHandler=null),this.pageHideHandler&&(window.removeEventListener("pagehide",this.pageHideHandler),this.pageHideHandler=null)},async resolveOpenUrl(){const e=++this.resolveGeneration,n=()=>e===this.resolveGeneration;if(this.isLoading=!0,this.loadError="",this.iframeSrc="",this.externalOpenUrl="",this.externalOpenMessage="",this.externalSnapshotText="",this.readonlySnapshotMode=!1,this.readonlySnapshotText="",this.readonlySnapshotHtml="",this.syncUrl="",this.syncInFlight=!1,this.stopSyncLoop(),!this.filePath){if(!n())return;this.loadError="No .pad file selected.",this.isLoading=!1;return}const s=L(),a=f("/apps/"+y+"/api/v1/pads/open"),i=(()=>{if(!s)return"";const l=new URL(f("/apps/"+y+"/api/v1/public/open/"+encodeURIComponent(s)),window.location.origin);return l.searchParams.set("file",this.filePath),l.toString()})(),h=this.resolvedFileId!==null?f("/apps/"+y+"/api/v1/pads/open-by-id"):"",d=new URLSearchParams;d.set("file",this.filePath);const c=new URLSearchParams;this.resolvedFileId!==null&&c.set("fileId",String(this.resolvedFileId));const u={"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:m()};try{const l=async()=>{let o=null;if(i)o=await this.fetchOpenPayload(i);else if(h)try{o=await this.fetchOpenPayload(h,{method:"POST",headers:u,body:c.toString()})}catch{}return o||(o=await this.fetchOpenPayload(a,{method:"POST",headers:u,body:d.toString()})),o};let t;try{if(t=await l(),!n())return}catch(o){if(!this.isMissingFrontmatterError(o))throw o;if(await this.initializeMissingFrontmatter(),!n()||(t=await l(),!n()))return}this.syncUrl=t&&typeof t.sync_url=="string"?t.sync_url:"";const r=Number(t&&t.sync_interval_seconds);if(this.syncIntervalMs=Number.isFinite(r)&&r>0?Math.max(5e3,Math.min(36e5,r*1e3)):12e4,this.installSyncLifecycleHandlers(),this.syncUrl&&this.startSyncLoop(),t&&t.is_readonly_snapshot===!0){this.readonlySnapshotMode=!0,this.readonlySnapshotText=typeof t.snapshot_text=="string"?t.snapshot_text:"",this.readonlySnapshotHtml=typeof t.snapshot_html=="string"?t.snapshot_html:"",this.markLoaded();return}if(t&&t.is_external===!0&&typeof t.url=="string"&&t.url.trim()!==""){const o=t.url.trim();this.externalOpenUrl=o,this.externalOpenMessage=p("This view shows the last synced snapshot stored in the .pad file. It is read-only here. To edit the pad, open the original pad in a new tab."),this.externalSnapshotText=t&&typeof t.snapshot_text=="string"?t.snapshot_text:"",this.markLoaded();return}this.iframeSrc=t.url,this.markLoaded()}catch(l){if(!n())return;this.loadError=l instanceof Error?l.message:"Could not load pad.",this.markLoaded()}finally{if(!n())return;this.isLoading=!1}}},beforeDestroy(){this.resolveGeneration+=1,this.runSync(!0,!0),this.stopSyncLoop(),this.removeSyncLifecycleHandlers()},beforeUnmount(){this.resolveGeneration+=1,this.runSync(!0,!0),this.stopSyncLoop(),this.removeSyncLifecycleHandlers()},render(e){return this.loadError?e("div",{class:"epnc-native-status epnc-native-status--error"},[e("div",{class:"epnc-native-error-card"},[e("div",{class:"epnc-native-error-title"},"Unable to open pad"),e("div",{class:"epnc-native-error-message"},this.loadError)])]):this.externalOpenUrl?e("div",{class:"epnc-native-status"},[e("div",{class:"epnc-native-error-card"},[e("div",{class:"epnc-native-error-title"},p("Pad from another server")),e("div",{class:"epnc-native-error-message"},this.externalOpenMessage),e("pre",{class:"epnc-native-preview"},this.externalSnapshotText.trim()!==""?this.externalSnapshotText:p("No synced snapshot is stored in this .pad file yet.")),e("a",{class:"button primary",attrs:{href:this.externalOpenUrl,target:"_blank",rel:"noopener noreferrer"}},p("Open original pad in new tab"))])]):this.readonlySnapshotMode?e("div",{class:"epnc-native-snapshot"},[e("div",{class:"epnc-native-snapshot__inner"},[e("div",{class:"epnc-native-snapshot__title"},p("Read-only snapshot")),e("div",{class:"epnc-native-snapshot__message"},p("This share shows the last synced snapshot stored in the .pad file.")),this.readonlySnapshotHtml.trim()!==""?e("div",{class:"epnc-native-snapshot__text epnc-native-snapshot__text--html",domProps:{innerHTML:this.readonlySnapshotHtml}}):e("pre",{class:"epnc-native-snapshot__text"},this.readonlySnapshotText.trim()!==""?this.readonlySnapshotText:p("No synced snapshot is stored in this .pad file yet."))])]):this.isLoading||!this.iframeSrc?e("div",{class:"epnc-native-status"},"Loading pad..."):e("div",{class:"epnc-native-shell"},[e("iframe",{attrs:{src:this.iframeSrc},on:{load:()=>this.markLoaded(),error:()=>this.markLoaded()},class:"epnc-native-iframe"})])}},w=()=>{if(v+=1,!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.registerHandler=="function")){v<20&&window.setTimeout(w,500);return}Array.isArray(window.OCA.Viewer.availableHandlers)&&window.OCA.Viewer.availableHandlers.some(e=>e&&e.id===g)||window.OCA.Viewer.registerHandler({id:g,mimes:[b],component:S})};w()})();
 //# sourceMappingURL=etherpad_nextcloud-viewer-main.mjs.map
