+++ js/oc-compat-C8UpNj_V.chunk.mjs
@@ -1,0 +1,2 @@
+const c="etherpad_nextcloud",u="application/x-etherpad-nextcloud",f="etherpad_nextcloud",C=n=>window.OC&&typeof window.OC.generateUrl=="function"?window.OC.generateUrl(n):"/index.php"+n,p=(n,o)=>window.OC&&typeof window.OC.imagePath=="function"?window.OC.imagePath(n,o):"",O=(n="")=>{const o=String(n||"").trim();return o!==""?o:String(window.OC&&window.OC.requestToken||"")},s=()=>{if(window.OC&&typeof window.OC.getCurrentUser=="function"){const n=window.OC.getCurrentUser();return String(n&&n.uid||"").trim()}return""},a=(n,o)=>{const e=s(),i=String(n||"").trim();if(e===""||i==="")return"";const t=window.OC&&typeof window.OC.linkToRemoteBase=="function"?window.OC.linkToRemoteBase("dav"):"/remote.php/dav",r=new URL(t,window.location.origin),d=i.split("/").filter(w=>w!=="").map(w=>w).join("/");return r.origin+r.pathname.replace(/\/+$/,"")+"/files/"+e+"/"+d},m=n=>a(n),g=(n,o)=>{const e=window._nc_event_bus||window.OC&&window.OC._eventBus;return!e||typeof e.emit!="function"?!1:(e.emit(n,o),!0)},l=()=>{const n=window.OC&&window.OC.PERMISSION_READ,o=Number(n);return Number.isFinite(o)&&o>0?o:1},_=()=>{const n=window.OC&&window.OC.PERMISSION_CREATE,o=Number(n);return Number.isFinite(o)&&o>0?o:4},h=()=>{const n=[window.OC&&window.OC.config&&window.OC.config.version,window.OC&&window.OC.config&&window.OC.config.versionstring,window.oc_appconfig&&window.oc_appconfig.core&&window.oc_appconfig.core.version,document.documentElement&&document.documentElement.getAttribute("data-version")];for(const o of n){const e=String(o||"").trim();if(e==="")continue;const i=e.match(/^(\d+)/);if(!i)continue;const t=parseInt(i[1],10);if(Number.isFinite(t)&&t>0)return t}return null},v=n=>typeof window.t=="function"?window.t(c,n):n;export{c as A,u as M,f as V,l as a,p as b,C as c,m as d,g as e,_ as f,h as n,O as o,v as t};
+//# sourceMappingURL=oc-compat-C8UpNj_V.chunk.mjs.map
